+++ java/integTest/src/test/java/glide/TransactionTestUtilities.java
@@ -16,6 +16,8 @@ public class TransactionTestUtilities {
     private static final String key4 = "{key}" + UUID.randomUUID();
     private static final String key5 = "{key}" + UUID.randomUUID();
     private static final String key6 = "{key}" + UUID.randomUUID();
+    // TODO rename after #160 merge & rebase
+    private static final String listKey3 = "{key}:listKey3-" + UUID.randomUUID();
     private static final String key7 = "{key}" + UUID.randomUUID();
     private static final String key8 = "{key}" + UUID.randomUUID();
     private static final String value1 = UUID.randomUUID().toString();
@@ -99,6 +101,9 @@ public class TransactionTestUtilities {
 
         baseTransaction.echo("GLIDE");
 
+        // TODO should be before LINDEX from #176 and BRPOP/BLPOP from #174
+        baseTransaction.rpushx(listKey3, new String[] {"_"}).lpushx(listKey3, new String[] {"_"});
+
         return baseTransaction;
     }
 
@@ -157,6 +162,8 @@ public class TransactionTestUtilities {
             Map.of("timeout", "1000"),
             OK,
             "GLIDE", // echo
+            0L, // rpushx(listKey3, new String[] { "_" })
+            0L, // lpushx(listKey3, new String[] { "_" })
         };
     }
 }